# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nfc locks api",
        "description": "Nfc locks",
        "contact": {
            "email": "tcom.developer@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://blog.test/api",
            "description": "local server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Prijava korisnika za web aplikaciju",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "ime korisnika",
                                        "type": "string",
                                        "default": "korisnik"
                                    },
                                    "password": {
                                        "description": "lozinka",
                                        "type": "string",
                                        "default": "korisnik"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Korisnik prijavljen"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Odjava korisnika",
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Korisnik odjavljen"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/whoami": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Podaci o korisniku na osnovu tokena",
                "operationId": "aa7bc66822f59c40cade93c5e4c3b41e",
                "responses": {
                    "200": {
                        "description": "Podaci o korisniku vraćeni na front"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/change-password": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Izmena password-a za korisnika",
                "operationId": "c219c3b643e1caf71a27e3ffa7e8eeaa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "old_password",
                                    "new_password"
                                ],
                                "properties": {
                                    "old_password": {
                                        "type": "string",
                                        "default": "Trenutna lozinka"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "default": "Nova lozinka",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Izmena password-a korisnika"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/config": {
            "get": {
                "tags": [
                    "Config"
                ],
                "summary": "Konfiguracioni fajl blog",
                "operationId": "107d362ad26109dd3ada26c0f348fce1",
                "responses": {
                    "200": {
                        "description": "Sve postojeće konfiguracije"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/config/{code}": {
            "get": {
                "tags": [
                    "Config"
                ],
                "summary": "Pojedinačni konfiguracioni podatak",
                "description": "order_models, order_names, config_substation_status, config_order_types, config_user_types, config_metalwork, config_order_status, config_signal_score, config_substation_type, config_nfc_lock_type",
                "operationId": "18d0bda0c6483c744d96e47258ec17a6",
                "parameters": [
                    {
                        "parameter": "code",
                        "name": "code",
                        "in": "path",
                        "description": "naziv pojeninačnog šifarnika",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pojedinačni konfiguracioni podatak"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lista korisnika",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter korisnika prema tipu korisnika",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "technician"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filter korisnika prema statusu aktivnosti",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista korisnika"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/show/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Korisnik",
                "operationId": "3d7733a19ba63f5d81ae3edf55d42b8b",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "Users id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Korisnik"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/store": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Kreiranje novog korisnika",
                "operationId": "f2340ca439313e43026cfa64efa7c769",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "default": "korisnicko ime"
                                    },
                                    "password": {
                                        "type": "string",
                                        "default": "12345678",
                                        "minLength": 8
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "default": "ime korisnika"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "default": "prezime korisnika"
                                    },
                                    "type": {
                                        "type": "string",
                                        "default": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kreiranje novog korisnika"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/update/{id}": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Izmena postojećeg korisnika",
                "operationId": "b29ea35945546eef517186e6f394e27d",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "Users id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "default": "korisnicko ime"
                                    },
                                    "password": {
                                        "type": "string",
                                        "default": "12345678",
                                        "minLength": 8
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "default": "ime korisnika"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "default": "prezime korisnika"
                                    },
                                    "type": {
                                        "type": "string",
                                        "default": "admin"
                                    },
                                    "active": {
                                        "type": "integer",
                                        "default": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kreiranje novog korisnika"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/delete/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Brisanje korisnika",
                "operationId": "672077db1c1a041839951757eb77f1ce",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "Users id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Korisnik je obrisan"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Controller": {
                "format": "https"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authorize with bearer token",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Users",
            "description": "Users"
        },
        {
            "name": "Config",
            "description": "Config"
        }
    ]
}